Moved the texts into a separate file

+++ task-06/friendly_neighbourhood_bot.py
@@ -3,6 +3,7 @@ from discord.ext import commands
 from dotenv import load_dotenv
 import asyncio
 import os
+import mytexts
 
 load_dotenv()
 token=os.getenv("DISCORD_TOKEN")
@@ -11,38 +12,20 @@ new_student_role_id=int(os.getenv("NEW_STUDENT_ROLE_ID"))
 annoucnement_channel_id=int(os.getenv("ANNOUNCEMENT_CHANNEL_ID"))
 Authorised_role=["Faculty","Administrator"]
 
-description ="Your friendly neighbourhood bot. You can call me 'Sting-Annan'"
-
 intents = discord.Intents.default()
 intents.members = True
 intents.message_content = True
 
 FORBIDDEN_KEYWORDS=["villainous spam","unauthorized link","off-topic disruption","menacing threats"]
 
-RULES="1. Respect Everyone: Treat your fellow students, faculty, and staff with respect. This means no harassment, bullying, or discrimination of any kind.\n\n" \
-    "2. Stay On-Topic: Use the designated channels for their intended purpose.\n\n" \
-    "3. No Spam or Self-Promotion: Avoid posting repetitive messages, excessive emojis, or unapproved links.\n\n" \
-    "4. Keep it SFW: This is a college server, so please keep all content and conversations appropriate for a general audience." \
-    "5. Use Common Sense: If you're unsure if something is allowed, it's probably best to proceed with caution. If a rule doesn't cover a specific situation, mods will use their discretion to ensure the server remains a positive environment.\n\n" \
-    "If you have any questions or concerns, please feel free to DM a mod. Thanks for helping us make this a great community!"
-
-RESOURCES="‣ Khan Academy: https://www.khanacademy.org/ \n\n" \
-    "‣ Coursera: https://www.coursera.org/ \n\n"\
-    "‣ Evernote: https://evernote.com/ \n\n"\
-    "‣ Codecademy: https://www.codecademy.com/ \n\n"\
-    "‣ Google Scholar: https://scholar.google.com/"
-
-CONTACTS="‣ Dr. Evelyn Reed, Dean of Students\n\t‣Email: ereed@midtown.edu \n\n"\
-    "‣ Prof. Alan Finch, Head of Computer Science Department\n\t‣Email: afinch@midtown.edu\n\n"\
-    "‣ Mr. Ben Carter, Career Services Director\n\t‣Email: bcarter@midtown.edu\n\n"\
-    "‣ Dr. Sophia Chen, Head Counselor, Health & Wellness\n\t‣Email: schen@midtown.edu"
-
 async def auto_delete(msg):
-    time_in_secs=60 #CHANGE TO CONTROL THE TIME UNTIL DELETION (24hr - 86400sec)
-    await asyncio.sleep(time_in_secs)
-    await msg.delete()
+    time_in_secs=30 #CHANGE TO CONTROL THE TIME UNTIL DELETION (24hr - 86400sec)
+    #print(msg.pinned)
+    if msg.pinned==False:
+        await asyncio.sleep(time_in_secs)
+        await msg.delete()
 
-bot = commands.Bot(command_prefix='??',description=description,intents=intents)
+bot = commands.Bot(command_prefix='??',description=mytexts.description,intents=intents)
 @bot.event
 async def on_ready():
     print(f"Logged in as {bot.user.name}.\nBot is now running")
@@ -94,13 +77,13 @@ async def bugle_error(ctx,error):
 @bot.command()
 async def wisdom(ctx,*,content:str):
     if content.lower()=="rules":
-        await ctx.send(RULES)
+        await ctx.send(mytexts.RULES)
     elif content.lower()=="resources":
-        await ctx.send(RESOURCES)
+        await ctx.send(mytexts.RESOURCES)
     elif content.lower()=="contacts":
-        await ctx.send(CONTACTS)
+        await ctx.send(mytexts.CONTACTS)
     else:
-        await ctx.send("Invalid command usage. Use '??help' for more information")
+        await ctx.send(f"{ctx.author.mention} - Invalid command usage. Use '??help' for more information")
 
 #NOTE: Update command descriptions
 #NOTE: Implement pinned msg persistance on #announcements
